+++ gnuradio-runtime/include/gnuradio/sync_interpolator.h
@@ -51,7 +51,7 @@ namespace gr {
     void set_interpolation(unsigned interpolation)
     {
       d_interpolation = interpolation;
-      set_relative_rate(interpolation, 1);
+      set_relative_rate((uint64_t) interpolation, 1);
       set_output_multiple(interpolation);
     }
