+++ gr-digital/lib/ofdm_chanest_vcvc_impl.cc
@@ -128,8 +128,8 @@ namespace gr {
       }
 
       set_output_multiple(d_n_data_syms);
-      set_relative_rate((unsigned) d_n_data_syms,
-                        (unsigned) (d_n_data_syms + d_n_sync_syms));
+      set_relative_rate((uint64_t) d_n_data_syms,
+                        (uint64_t) (d_n_data_syms + d_n_sync_syms));
       set_tag_propagation_policy(TPP_DONT);
     }
